Move functionality into `BuildInScalarFunction` (#6612)
