+++ datafusion/expr/src/lib.rs
@@ -37,7 +37,6 @@ pub mod expr_rewriter;
 pub mod expr_schema;
 pub mod field_util;
 pub mod function;
-pub mod function_err;
 mod literal;
 pub mod logical_plan;
 mod nullif;
